Refactor bot structure
