+++ db_access.py
@@ -1,0 +1,39 @@
+from datetime import datetime
+from typing import List
+from db import SessionLocal, Profile, Entry
+
+
+def save_profile(user_id: int, icr: float, cf: float, target: float) -> None:
+    with SessionLocal() as session:
+        profile = session.get(Profile, user_id)
+        if not profile:
+            profile = Profile(telegram_id=user_id)
+            session.add(profile)
+        profile.icr = icr
+        profile.cf = cf
+        profile.target_bg = target
+        session.commit()
+
+
+def get_profile(user_id: int) -> Profile | None:
+    with SessionLocal() as session:
+        return session.get(Profile, user_id)
+
+
+def add_entry(entry_data: dict) -> None:
+    session = SessionLocal()
+    entry = Entry(**entry_data)
+    session.add(entry)
+    session.commit()
+    session.close()
+
+
+def get_entries_since(user_id: int, date_from: datetime) -> List[Entry]:
+    with SessionLocal() as s:
+        return (
+            s.query(Entry)
+            .filter(Entry.telegram_id == user_id)
+            .filter(Entry.event_time >= date_from)
+            .order_by(Entry.event_time)
+            .all()
+        )
